Changing user profile images functionality & mayor changes/fixes
The users can now change their profile image. After uploading the old one gets deleted of the S3 bucket and the new one is added to the S3 bucket.

Added folders to the S3 bucket. One folder 'profile-images' for the profile pictures and the 'location-images' for the location images.

I have also restructured and rewritten the  's3-bucket' service, so it is more easily readable and to better learn and understand the code.