Forgot password functionality better and safer
When the user requests a password reset token, we check if the user even exists. And no matter the outcome we always display the same message of 'email has been sent' as to not to tell them which emails are registered at our site. If the user does exist, we generate a 64 char string as a token, hash it in SHA256 and store it to the database, set an expiration date of only 2 hours and send the user an email with the link to the change form site.  The form on the site follows the same password rules as the sign in